fix: [2.16] rename model_registry to ai_hub
